Extract dispatcher, supervisor, bridge, entrypoint, log pump, paths, and workspaces (#82591)
