+++ hermes_cli/kanban_runtime/scaffolding.py
@@ -1,0 +1,17 @@
+"""Graph-gated owner generated by the #82591 campaign."""
+
+from typing import Any
+
+_EPHEMERAL_SCAFFOLDING_FLAGS = (
+    "__scaffolding",
+    "__recovery",
+)
+
+
+
+def _is_ephemeral_scaffolding(msg: Any) -> bool:
+    """Return True when ``msg`` is internal recovery scaffolding that must never
+    be persisted to the durable transcript (SQLite session store or JSON log)."""
+    return isinstance(msg, dict) and any(
+        msg.get(flag) for flag in _EPHEMERAL_SCAFFOLDING_FLAGS
+    )
